// frontend (2)/src/app/types/dto/person/related/contact/phone/person-phone-number.dto.ts
import { PhoneNumberDto } from '../../../../phone-number.dto';
import { PersonPhoneNumberDtoId } from '../../reference/contact/person-phone-number-dto-id';


export interface PersonPhoneNumberDto {

    id: PersonPhoneNumberDtoId

    phoneNumber: PhoneNumberDto;

    mainPhoneNumber: boolean;

}